requirements.txt
psycopg2-binary==2.9.10
beautifulsoup4==4.13.4
requests==2.32.3
pandas==2.2.3
pyarrow==20.0.0
python-dateutil==2.9.0.post0